+++ lib/src/widget.dart
@@ -225,7 +225,10 @@ class FlexiChip extends StatelessWidget {
   /// for the sake of extensions that add support for additional events.
   final FlexiChipEventController? eventsController;
 
+  /// The curve to apply when animating the parameters of this widget.
   final Curve curve;
+
+  /// The duration over which to animate the parameters of this widget.
   final Duration duration;
 
   static const defaultDuration = Duration(milliseconds: 200);
@@ -301,201 +304,27 @@ class _FlexiChipRender extends ImplicitlyAnimatedWidget {
           curve: curve,
         );
 
-  /// The primary content of the chip.
-  ///
-  /// Typically a [Text] widget.
   final Widget label;
-
-  /// Typically used as profile image.
-  ///
-  /// If the avatar is to have the user's initials, use [avatarText] instead.
   final ImageProvider? avatarImage;
-
-  /// The primary content of the chip avatar.
-  ///
-  /// Typically a [Text] widget.
   final Widget? avatarText;
-
-  /// A custom widget to display prior to the chip's [label].
   final Widget? leading;
-
-  /// A custom widget to display next to the chip's [label].
   final Widget? trailing;
-
-  /// Tooltip string to be used for the body area (where the label and avatar
-  /// are) of the chip.
   final String? tooltip;
-
-  /// The icon displayed when [onDeleted] is set.
-  ///
-  /// Defaults to an [Icon] widget set to use [Icons.cancel].
   final Widget? deleteIcon;
-
-  /// The message to be used for the chip's delete button tooltip.
-  ///
-  /// If provided with an empty string, the tooltip of the delete button will be
-  /// disabled.
-  ///
-  /// If null, the default [MaterialLocalizations.deleteButtonTooltip] will be
-  /// used.
   final String? deleteTooltip;
-
-  /// Whether or not this chip is selected.
-  ///
-  /// If [onSelected] is not null, this value will be used to determine if the
-  /// select checkmark will be shown or not.
-  ///
-  /// Must not be null. Defaults to false.
   final bool selected;
-
-  /// Whether or not this chip is disabled for input.
-  ///
-  /// Defaults to false. Cannot be null.
   final bool disabled;
-
-  /// Whether or not to show a checkmark when [selected] is true.
-  ///
-  /// Defaults to false. Cannot ve null.
   final bool checkmark;
-
-  /// True if this widget will be selected as the initial focus
-  /// when no other node in its scope is currently focused.
-  ///
-  /// Ideally, there is only one widget with autofocus set in each [FocusScope].
-  /// If there is more than one widget with autofocus set,
-  /// then the first one added to the tree will get focus.
-  ///
-  /// Must not be null. Defaults to false.
   final bool autofocus;
-
-  /// An optional focus node to use as the focus node for this widget.
-  ///
-  /// If one is not supplied, then one will be automatically allocated, owned,
-  /// and managed by this widget. The widget will be focusable even if a [focusNode] is not supplied.
-  /// If supplied, the given focusNode will be hosted by this widget, but not owned.
-  /// See [FocusNode] for more information on what being hosted and/or owned implies.
-  ///
-  /// Supplying a focus node is sometimes useful if an ancestor
-  /// to this widget wants to control when this widget has the focus.
-  /// The owner will be responsible for calling [FocusNode.dispose]
-  /// on the focus node when it is done with it, but this widget
-  /// will attach/detach and reparent the node when needed.
   final FocusNode? focusNode;
-
-  /// The splash color of the ink response. If this property is null then the
-  /// splash color of the theme, [ThemeData.splashColor], will be used.
   final Color? splashColor;
-
-  /// Defines the appearance of the splash.
-  ///
-  /// Defaults to the value of the theme's splash factory: [ThemeData.splashFactory].
   final InteractiveInkFeatureFactory? splashFactory;
-
-  /// Called when the user taps the chip.
-  ///
-  /// If [onPressed] is set, then this callback will be called when the user
-  /// taps on the label or avatar parts of the chip. If [onPressed] is null,
-  /// then the chip will be disabled.
-  ///
-  /// {@tool snippet}
-  ///
-  /// ```dart
-  /// class Blacksmith extends StatelessWidget {
-  ///   const Blacksmith({Key? key}) : super(key: key);
-  ///
-  ///   void startHammering() {
-  ///     print('bang bang bang');
-  ///   }
-  ///
-  ///   @override
-  ///   Widget build(BuildContext context) {
-  ///     return FlexiChip(
-  ///       label: const Text('Apply Hammer'),
-  ///       onPressed: startHammering,
-  ///     );
-  ///   }
-  /// }
-  /// ```
-  /// {@end-tool}
   final VoidCallback? onPressed;
-
-  /// Called when the user taps the [deleteIcon] to delete the chip.
-  ///
-  /// If null, the delete button will not appear on the chip.
-  ///
-  /// The chip will not automatically remove itself: this just tells the app
-  /// that the user tapped the delete button.
   final VoidCallback? onDeleted;
-
-  /// Called when the chip should change between selected and de-selected
-  /// states.
-  ///
-  /// When the chip is tapped, then the [onSelected] callback, if set, will be
-  /// applied to `!selected` (see [selected]).
-  ///
-  /// The chip passes the new value to the callback but does not actually
-  /// change state until the parent widget rebuilds the chip with the new
-  /// value.
-  ///
-  /// The callback provided to [onSelected] should update the state of the
-  /// parent [StatefulWidget] using the [State.setState] method, so that the
-  /// parent gets rebuilt.
-  ///
-  /// The [onSelected] and [onPressed] callbacks must not
-  /// both be specified at the same time.
-  ///
-  /// {@tool snippet}
-  ///
-  /// A [StatefulWidget] that illustrates use of onSelected in an [InputChip].
-  ///
-  /// ```dart
-  /// class Wood extends StatefulWidget {
-  ///   const Wood({Key? key}) : super(key: key);
-  ///
-  ///   @override
-  ///   State<StatefulWidget> createState() => WoodState();
-  /// }
-  ///
-  /// class WoodState extends State<Wood> {
-  ///   bool _useChisel = false;
-  ///
-  ///   @override
-  ///   Widget build(BuildContext context) {
-  ///     return FlexiChip(
-  ///       label: const Text('Use Chisel'),
-  ///       selected: _useChisel,
-  ///       onSelected: (bool newValue) {
-  ///         setState(() {
-  ///           _useChisel = newValue;
-  ///         });
-  ///       },
-  ///     );
-  ///   }
-  /// }
-  /// ```
-  /// {@end-tool}
   final ValueChanged<bool>? onSelected;
-
-  /// The style to be applied to the chip.
-  ///
-  /// If [style] is an event driven [FlexiChipStyle]
-  /// by [FlexiChipStyle.driven], then [FlexiChipStyle.evaluate]
-  /// is used for the following [FlexiChipEvent]s:
-  ///
-  ///  * [FlexiChipEvent.disabled].
-  ///  * [FlexiChipEvent.selected].
-  ///  * [FlexiChipEvent.hovered].
-  ///  * [FlexiChipEvent.focused].
-  ///  * [FlexiChipEvent.pressed].
   final FlexiChipStyle? style;
-
-  /// Used by widgets that expose their internal event
-  /// for the sake of extensions that add support for additional events.
   final FlexiChipEventController? eventsController;
-
   final ThemeData theme;
-
-  // static const defaultDuration = Duration(milliseconds: 200);
 
   bool get enabled => !disabled;
 
@@ -513,8 +342,6 @@ class _FlexiChipRender extends ImplicitlyAnimatedWidget {
 
 class _FlexiChipRenderState extends AnimatedWidgetBaseState<_FlexiChipRender>
     with WidgetEventMixin<_FlexiChipRender> {
-  // ThemeData appTheme = ThemeData();
-
   FlexiChipStyle style = const FlexiChipStyle();
 
   bool get isFilled => style.backgroundOpacity == 1;
@@ -614,6 +441,10 @@ class _FlexiChipRenderState extends AnimatedWidgetBaseState<_FlexiChipRender>
     );
   }
 
+  EdgeInsetsGeometry get containerMargin {
+    return style.margin ?? FlexiChipStyle.defaultMargin;
+  }
+
   BorderRadiusGeometry get containerRadius {
     return style.borderRadius ?? FlexiChipStyle.defaultBorderRadius;
   }
@@ -632,18 +463,6 @@ class _FlexiChipRenderState extends AnimatedWidgetBaseState<_FlexiChipRender>
       ),
     );
   }
-
-  // BoxDecoration get containerDecoration {
-  //   return BoxDecoration(
-  //     color: backgroundColor,
-  //     border: Border.all(
-  //       color: borderColor,
-  //       width: style.borderWidth ?? FlexiChipStyle.defaultBorderWidth,
-  //       style: style.borderStyle ?? FlexiChipStyle.defaultBorderStyle,
-  //     ),
-  //     borderRadius: style.borderRadius,
-  //   );
-  // }
 
   TextStyle get foregroundStyle {
     return const TextStyle()
@@ -745,15 +564,14 @@ class _FlexiChipRenderState extends AnimatedWidgetBaseState<_FlexiChipRender>
     return _containerBorderTween?.evaluate(animation) ?? containerBorder;
   }
 
-  // DecorationTween? _containerDecorationTween;
-  // Decoration get animatedContainerDecoration {
-  //   return _containerDecorationTween?.evaluate(animation) ??
-  //       containerDecoration;
-  // }
-
   EdgeInsetsGeometryTween? _containerPaddingTween;
   EdgeInsetsGeometry get animatedContainerPadding {
     return _containerPaddingTween?.evaluate(animation) ?? containerPadding;
+  }
+
+  EdgeInsetsGeometryTween? _containerMarginTween;
+  EdgeInsetsGeometry get animatedContainerMargin {
+    return _containerMarginTween?.evaluate(animation) ?? containerMargin;
   }
 
   ColorTween? _foregroundColorTween;
@@ -914,38 +732,30 @@ class _FlexiChipRenderState extends AnimatedWidgetBaseState<_FlexiChipRender>
     widgetEvents.toggle(FlexiChipEvent.focused, value);
   }
 
-  // setTheme() {
-  //   if (mounted) appTheme = Theme.of(context);
-  // }
-
   @override
   void initState() {
-    WidgetsBinding.instance.addPostFrameCallback((_) {
-      initWidgetEvents(widget.eventsController);
-      widgetEvents.toggle(FlexiChipEvent.disabled, widget.disabled);
-      widgetEvents.toggle(FlexiChipEvent.selected, widget.selected);
-      // setTheme();
-      setStyle();
-    });
+    initWidgetEvents(widget.eventsController);
+    widgetEvents.toggle(FlexiChipEvent.disabled, widget.disabled);
+    widgetEvents.toggle(FlexiChipEvent.selected, widget.selected);
+    setStyle();
     super.initState();
   }
 
   @override
   void didChangeDependencies() {
-    WidgetsBinding.instance.addPostFrameCallback((_) {
-      // setTheme();
-      super.didChangeDependencies();
-      didUpdateWidget(widget);
-    });
+    super.didChangeDependencies();
+    didUpdateWidget(widget);
   }
 
   @override
   void didUpdateWidget(_FlexiChipRender oldWidget) {
-    updateWidgetEvents(oldWidget.eventsController, widget.eventsController);
-    widgetEvents.toggle(FlexiChipEvent.disabled, widget.disabled);
-    widgetEvents.toggle(FlexiChipEvent.selected, widget.selected);
-    setStyle();
-    super.didUpdateWidget(oldWidget);
+    if (mounted) {
+      updateWidgetEvents(oldWidget.eventsController, widget.eventsController);
+      widgetEvents.toggle(FlexiChipEvent.disabled, widget.disabled);
+      widgetEvents.toggle(FlexiChipEvent.selected, widget.selected);
+      setStyle();
+      super.didUpdateWidget(oldWidget);
+    }
   }
 
   @override
@@ -965,6 +775,12 @@ class _FlexiChipRenderState extends AnimatedWidgetBaseState<_FlexiChipRender>
     _containerPaddingTween = visitor(
       _containerPaddingTween,
       containerPadding,
+      (value) => EdgeInsetsGeometryTween(begin: value),
+    ) as EdgeInsetsGeometryTween?;
+
+    _containerMarginTween = visitor(
+      _containerMarginTween,
+      containerMargin,
       (value) => EdgeInsetsGeometryTween(begin: value),
     ) as EdgeInsetsGeometryTween?;
 
@@ -1112,37 +928,40 @@ class _FlexiChipRenderState extends AnimatedWidgetBaseState<_FlexiChipRender>
       child: _Tooltip(
         message: widget.tooltip,
         enabled: widget.canTap,
-        child: ChipContainer(
-          color: animatedContainerColor,
-          clipBehavior: containerClipBehavior,
-          shape: animatedContainerBorder,
-          shadowColor: animatedContainerShadowColor,
-          elevation: animatedContainerElevation,
-          height: animatedContainerHeight,
-          child: _ChipEvent(
-            enabled: widget.canTap,
-            autofocus: widget.autofocus,
-            focusNode: widget.focusNode,
-            splashFactory: widget.splashFactory,
-            overlayColor: animatedContainerOverlayColor,
-            splashColor: widget.splashColor,
-            onTap: onTap,
-            onTapCancel: onTapCancel,
-            onTapDown: onTapDown,
-            onHover: onHover,
-            onFocus: onFocus,
-            child: _ChipForeground(
-              textStyle: animatedForegroundStyle,
-              iconTheme: IconThemeData(
-                color: animatedIconColor,
-                size: animatedIconSize,
-                opacity: animatedIconOpacity,
+        child: Padding(
+          padding: animatedContainerMargin,
+          child: ChipContainer(
+            color: animatedContainerColor,
+            clipBehavior: containerClipBehavior,
+            shape: animatedContainerBorder,
+            shadowColor: animatedContainerShadowColor,
+            elevation: animatedContainerElevation,
+            height: animatedContainerHeight,
+            child: _ChipEvent(
+              enabled: widget.canTap,
+              autofocus: widget.autofocus,
+              focusNode: widget.focusNode,
+              splashFactory: widget.splashFactory,
+              overlayColor: animatedContainerOverlayColor,
+              splashColor: widget.splashColor,
+              onTap: onTap,
+              onTapCancel: onTapCancel,
+              onTapDown: onTapDown,
+              onHover: onHover,
+              onFocus: onFocus,
+              child: _ChipForeground(
+                textStyle: animatedForegroundStyle,
+                iconTheme: IconThemeData(
+                  color: animatedIconColor,
+                  size: animatedIconSize,
+                  opacity: animatedIconOpacity,
+                ),
+                padding: animatedContainerPadding,
+                spacing: animatedForegroundSpacing,
+                label: label,
+                leading: leading,
+                trailing: trailing,
               ),
-              padding: animatedContainerPadding,
-              spacing: animatedForegroundSpacing,
-              label: label,
-              leading: leading,
-              trailing: trailing,
             ),
           ),
         ),